@@ -1,5 +1,7 @@
 import requests
 
+import fsspec
+
 from ..spec import AbstractFileSystem
 from ..utils import infer_storage_options
 from .memory import MemoryFile
@@ -36,8 +38,11 @@ class GithubFileSystem(AbstractFileSystem):
     url = "https://api.github.com/repos/{org}/{repo}/git/trees/{sha}"
     rurl = "https://raw.githubusercontent.com/{org}/{repo}/{sha}/{path}"
     protocol = "github"
+    timeout = (60, 60)  # connect, read timeouts
 
-    def __init__(self, org, repo, sha=None, username=None, token=None, **kwargs):
+    def __init__(
+        self, org, repo, sha=None, username=None, token=None, timeout=None, **kwargs
+    ):
         super().__init__(**kwargs)
         self.org = org
         self.repo = repo
@@ -45,10 +50,14 @@ class GithubFileSystem(AbstractFileSystem):
             raise ValueError("Auth required both username and token")
         self.username = username
         self.token = token
+        if timeout is not None:
+            self.timeout = timeout
         if sha is None:
             # look up default branch (not necessarily "master")
             u = "https://api.github.com/repos/{org}/{repo}"
-            r = requests.get(u.format(org=org, repo=repo), **self.kw)
+            r = requests.get(
+                u.format(org=org, repo=repo), timeout=self.timeout, **self.kw
+            )
             r.raise_for_status()
             sha = r.json()["default_branch"]
 
@@ -79,9 +88,8 @@ class GithubFileSystem(AbstractFileSystem):
         List of string
         """
         r = requests.get(
-            "https://api.github.com/{part}/{org}/repos".format(
-                part=["users", "orgs"][is_org], org=org_or_user
-            )
+            f"https://api.github.com/{['users', 'orgs'][is_org]}/{org_or_user}/repos",
+            timeout=cls.timeout,
         )
         r.raise_for_status()
         return [repo["name"] for repo in r.json()]
@@ -90,8 +98,8 @@ class GithubFileSystem(AbstractFileSystem):
     def tags(self):
         """Names of tags in the repo"""
         r = requests.get(
-            "https://api.github.com/repos/{org}/{repo}/tags"
-            "".format(org=self.org, repo=self.repo),
+            f"https://api.github.com/repos/{self.org}/{self.repo}/tags",
+            timeout=self.timeout,
             **self.kw,
         )
         r.raise_for_status()
@@ -101,8 +109,8 @@ class GithubFileSystem(AbstractFileSystem):
     def branches(self):
         """Names of branches in the repo"""
         r = requests.get(
-            "https://api.github.com/repos/{org}/{repo}/branches"
-            "".format(org=self.org, repo=self.repo),
+            f"https://api.github.com/repos/{self.org}/{self.repo}/branches",
+            timeout=self.timeout,
             **self.kw,
         )
         r.raise_for_status()
@@ -151,7 +159,9 @@ class GithubFileSystem(AbstractFileSystem):
                 _sha = out["sha"]
         if path not in self.dircache or sha not in [self.root, None]:
             r = requests.get(
-                self.url.format(org=self.org, repo=self.repo, sha=_sha), **self.kw
+                self.url.format(org=self.org, repo=self.repo, sha=_sha),
+                timeout=self.timeout,
+                **self.kw,
             )
             if r.status_code == 404:
                 raise FileNotFoundError(path)
@@ -212,8 +222,18 @@ class GithubFileSystem(AbstractFileSystem):
         url = self.rurl.format(
             org=self.org, repo=self.repo, path=path, sha=sha or self.root
         )
-        r = requests.get(url, **self.kw)
+        r = requests.get(url, timeout=self.timeout, **self.kw)
         if r.status_code == 404:
             raise FileNotFoundError(path)
         r.raise_for_status()
         return MemoryFile(None, None, r.content)
+
+    def cat(self, path, recursive=False, on_error="raise", **kwargs):
+        paths = self.expand_path(path, recursive=recursive)
+        urls = [
+            self.rurl.format(org=self.org, repo=self.repo, path=u, sha=self.root)
+            for u, sh in paths
+        ]
+        fs = fsspec.filesystem("http")
+        data = fs.cat(urls, on_error="return")
+        return {u: v for ((k, v), u) in zip(data.items(), urls)}
