@@ -1,21 +1,38 @@
 import contextlib
+import gzip
 import json
 import os
 import threading
 from collections import ChainMap
 from http.server import BaseHTTPRequestHandler, HTTPServer
+from types import SimpleNamespace
 
 import pytest
 
 requests = pytest.importorskip("requests")
-port = 9898
 data = b"\n".join([b"some test data"] * 1000)
-realfile = "http://127.0.0.1:%i/index/realfile" % port
-index = b'<a href="%s">Link</a>' % realfile.encode()
 listing = open(
     os.path.join(os.path.dirname(__file__), "data", "listing.html"), "rb"
 ).read()
 win = os.name == "nt"
+
+
+def _make_realfile(baseurl):
+    return f"{baseurl}/index/realfile"
+
+
+def _make_index_listing(baseurl):
+    realfile = _make_realfile(baseurl)
+    return b'<a href="%s">Link</a>' % realfile.encode()
+
+
+def _make_listing(*paths):
+    def _make_listing_port(baseurl):
+        return "\n".join(
+            f'<a href="{baseurl}{f}">Link_{i}</a>' for i, f in enumerate(paths)
+        ).encode()
+
+    return _make_listing_port
 
 
 @pytest.fixture
@@ -31,8 +48,12 @@ class HTTPTestHandler(BaseHTTPRequestHandler):
     static_files = {
         "/index/realfile": data,
         "/index/otherfile": data,
-        "/index": index,
+        "/index": _make_index_listing,
         "/data/20020401": listing,
+        "/simple/": _make_listing("/simple/file", "/simple/dir/"),
+        "/simple/file": data,
+        "/simple/dir/": _make_listing("/simple/dir/file"),
+        "/simple/dir/file": data,
     }
     dynamic_files = {}
 
@@ -52,15 +73,23 @@ class HTTPTestHandler(BaseHTTPRequestHandler):
             self.wfile.write(data)
 
     def do_GET(self):
-        file_path = self.path.rstrip("/")
+        baseurl = f"http://127.0.0.1:{self.server.server_port}"
+        file_path = self.path
+        if file_path.endswith("/") and file_path.rstrip("/") in self.files:
+            file_path = file_path.rstrip("/")
         file_data = self.files.get(file_path)
+        if callable(file_data):
+            file_data = file_data(baseurl)
         if "give_path" in self.headers:
             return self._respond(200, data=json.dumps({"path": self.path}).encode())
+        if "redirect" in self.headers and file_path != "/index/realfile":
+            new_url = _make_realfile(baseurl)
+            return self._respond(301, {"Location": new_url})
         if file_data is None:
             return self._respond(404)
 
         status = 200
-        content_range = "bytes 0-%i/%i" % (len(file_data) - 1, len(file_data))
+        content_range = f"bytes 0-{len(file_data) - 1}/{len(file_data)}"
         if ("Range" in self.headers) and ("ignore_range" not in self.headers):
             ran = self.headers["Range"]
             b, ran = ran.split("=")
@@ -71,15 +100,26 @@ class HTTPTestHandler(BaseHTTPRequestHandler):
             else:
                 # suffix only
                 l = len(file_data)
-                content_range = f"bytes {l-int(end)}-{l-1}/{l}"
+                content_range = f"bytes {l - int(end)}-{l - 1}/{l}"
                 file_data = file_data[-int(end) :]
             if "use_206" in self.headers:
                 status = 206
         if "give_length" in self.headers:
-            response_headers = {"Content-Length": len(file_data)}
+            if "gzip_encoding" in self.headers:
+                file_data = gzip.compress(file_data)
+                response_headers = {
+                    "Content-Length": len(file_data),
+                    "Content-Encoding": "gzip",
+                }
+            else:
+                response_headers = {"Content-Length": len(file_data)}
             self._respond(status, response_headers, file_data)
         elif "give_range" in self.headers:
             self._respond(status, {"Content-Range": content_range}, file_data)
+        elif "give_mimetype" in self.headers:
+            self._respond(
+                status, {"Content-Type": "text/html; charset=utf-8"}, file_data
+            )
         else:
             self._respond(status, data=file_data)
 
@@ -90,7 +130,7 @@ class HTTPTestHandler(BaseHTTPRequestHandler):
             assert self.headers.get("Transfer-Encoding") == "chunked"
             self.files[file_path] = b"".join(self.read_chunks())
         else:
-            self.files[file_path] = self.rfile.read(length)
+            self.files[file_path] = self.rfile.read(int(length))
         self._respond(200)
 
     do_PUT = do_POST
@@ -107,10 +147,10 @@ class HTTPTestHandler(BaseHTTPRequestHandler):
             self.rfile.readline()
 
     def do_HEAD(self):
+        r_headers = {}
         if "head_not_auth" in self.headers:
-            return self._respond(
-                403, {"Content-Length": 123}, b"not authorized for HEAD request"
-            )
+            r_headers["Content-Length"] = 123
+            return self._respond(403, r_headers, b"not authorized for HEAD request")
         elif "head_ok" not in self.headers:
             return self._respond(405)
 
@@ -120,30 +160,34 @@ class HTTPTestHandler(BaseHTTPRequestHandler):
             return self._respond(404)
 
         if ("give_length" in self.headers) or ("head_give_length" in self.headers):
-            response_headers = {"Content-Length": len(file_data)}
             if "zero_length" in self.headers:
-                response_headers["Content-Length"] = 0
-
-            self._respond(200, response_headers)
+                r_headers["Content-Length"] = 0
+            elif "gzip_encoding" in self.headers:
+                file_data = gzip.compress(file_data)
+                r_headers["Content-Encoding"] = "gzip"
+                r_headers["Content-Length"] = len(file_data)
+            else:
+                r_headers["Content-Length"] = len(file_data)
         elif "give_range" in self.headers:
-            self._respond(
-                200, {"Content-Range": "0-%i/%i" % (len(file_data) - 1, len(file_data))}
-            )
+            r_headers["Content-Range"] = f"0-{len(file_data) - 1}/{len(file_data)}"
         elif "give_etag" in self.headers:
-            self._respond(200, {"ETag": "xxx"})
-        else:
-            self._respond(200)  # OK response, but no useful info
+            r_headers["ETag"] = "xxx"
+
+        if self.headers.get("accept_range") == "none":
+            r_headers["Accept-Ranges"] = "none"
+
+        self._respond(200, r_headers)
 
 
 @contextlib.contextmanager
 def serve():
-    server_address = ("", port)
+    server_address = ("", 0)
     httpd = HTTPServer(server_address, HTTPTestHandler)
     th = threading.Thread(target=httpd.serve_forever)
     th.daemon = True
     th.start()
     try:
-        yield "http://127.0.0.1:%i" % port
+        yield f"http://127.0.0.1:{httpd.server_port}"
     finally:
         httpd.socket.close()
         httpd.shutdown()
@@ -153,4 +197,5 @@ def serve():
 @pytest.fixture(scope="module")
 def server():
     with serve() as s:
-        yield s
+        server = SimpleNamespace(address=s, realfile=_make_realfile(s))
+        yield server
