@@ -1,16 +1,10 @@
-try:
-    from importlib.metadata import entry_points
-except ImportError:  # python < 3.8
-    try:
-        from importlib_metadata import entry_points
-    except ImportError:
-        entry_points = None
+from importlib.metadata import entry_points
 
-
-from . import _version, caching
+from . import caching
+from ._version import __version__  # noqa: F401
 from .callbacks import Callback
 from .compression import available_compressions
-from .core import get_fs_token_paths, open, open_files, open_local
+from .core import get_fs_token_paths, open, open_files, open_local, url_to_fs
 from .exceptions import FSTimeoutError
 from .mapping import FSMap, get_mapper
 from .registry import (
@@ -21,8 +15,6 @@ from .registry import (
     registry,
 )
 from .spec import AbstractFileSystem
-
-__version__ = _version.get_versions()["version"]
 
 __all__ = [
     "AbstractFileSystem",
@@ -41,6 +33,7 @@ __all__ = [
     "Callback",
     "available_protocols",
     "available_compressions",
+    "url_to_fs",
 ]
 
 
@@ -55,10 +48,21 @@ def process_entries():
                 specs = eps.select(group="fsspec.specs")
             else:
                 specs = eps.get("fsspec.specs", [])
+            registered_names = {}
             for spec in specs:
                 err_msg = f"Unable to load filesystem from {spec}"
+                name = spec.name
+                if name in registered_names:
+                    continue
+                registered_names[name] = True
                 register_implementation(
-                    spec.name, spec.value.replace(":", "."), errtxt=err_msg
+                    name,
+                    spec.value.replace(":", "."),
+                    errtxt=err_msg,
+                    # We take our implementations as the ones to overload with if
+                    # for some reason we encounter some, may be the same, already
+                    # registered
+                    clobber=True,
                 )
 
 
