@@ -1,6 +1,7 @@
 """
 fsspec user-defined exception classes
 """
+
 import asyncio
 
 
@@ -10,12 +11,8 @@ class BlocksizeMismatchError(ValueError):
     written with
     """
 
-    ...
-
 
 class FSTimeoutError(asyncio.TimeoutError):
     """
     Raised when a fsspec function timed out occurs
     """
-
-    ...
