@@ -20,11 +20,10 @@ def cli(tmpdir):
     try:
         yield client
     finally:
-        client.close()
+        client.shutdown()
 
 
 def test_basic(cli):
-
     fs = fsspec.filesystem("dask", target_protocol="memory")
     assert fs.ls("", detail=False) == ["/afile"]
     assert fs.cat("/afile") == b"data"
