@@ -1,5 +1,4 @@
-from __future__ import absolute_import, division, print_function
-
+import os
 import zipfile
 
 import fsspec
@@ -50,10 +49,15 @@ class ZipFileSystem(AbstractArchiveFileSystem):
         if mode not in set("rwa"):
             raise ValueError(f"mode '{mode}' no understood")
         self.mode = mode
-        if isinstance(fo, str):
+        if isinstance(fo, (str, os.PathLike)):
+            if mode == "a":
+                m = "r+b"
+            else:
+                m = mode + "b"
             fo = fsspec.open(
-                fo, mode=mode + "b", protocol=target_protocol, **(target_options or {})
+                fo, mode=m, protocol=target_protocol, **(target_options or {})
             )
+        self.force_zip_64 = allowZip64
         self.of = fo
         self.fo = fo.__enter__()  # the whole instance is a context
         self.zip = zipfile.ZipFile(
@@ -73,7 +77,7 @@ class ZipFileSystem(AbstractArchiveFileSystem):
     def __del__(self):
         if hasattr(self, "zip"):
             self.close()
-        del self.zip
+            del self.zip
 
     def close(self):
         """Commits any write changes to the file. Done on ``del`` too."""
@@ -85,14 +89,18 @@ class ZipFileSystem(AbstractArchiveFileSystem):
             # not read from the file.
             files = self.zip.infolist()
             self.dir_cache = {
-                dirname + "/": {"name": dirname + "/", "size": 0, "type": "directory"}
+                dirname.rstrip("/"): {
+                    "name": dirname.rstrip("/"),
+                    "size": 0,
+                    "type": "directory",
+                }
                 for dirname in self._all_dirnames(self.zip.namelist())
             }
             for z in files:
                 f = {s: getattr(z, s, None) for s in zipfile.ZipInfo.__slots__}
                 f.update(
                     {
-                        "name": z.filename,
+                        "name": z.filename.rstrip("/"),
                         "size": z.file_size,
                         "type": ("directory" if z.is_dir() else "file"),
                     }
@@ -115,13 +123,55 @@ class ZipFileSystem(AbstractArchiveFileSystem):
         path = self._strip_protocol(path)
         if "r" in mode and self.mode in set("wa"):
             if self.exists(path):
-                raise IOError("ZipFS can only be open for reading or writing, not both")
+                raise OSError("ZipFS can only be open for reading or writing, not both")
             raise FileNotFoundError(path)
         if "r" in self.mode and "w" in mode:
-            raise IOError("ZipFS can only be open for reading or writing, not both")
-        out = self.zip.open(path, mode.strip("b"))
+            raise OSError("ZipFS can only be open for reading or writing, not both")
+        out = self.zip.open(path, mode.strip("b"), force_zip64=self.force_zip_64)
         if "r" in mode:
             info = self.info(path)
             out.size = info["size"]
             out.name = info["name"]
         return out
+
+    def find(self, path, maxdepth=None, withdirs=False, detail=False, **kwargs):
+        if maxdepth is not None and maxdepth < 1:
+            raise ValueError("maxdepth must be at least 1")
+
+        # Remove the leading slash, as the zip file paths are always
+        # given without a leading slash
+        path = path.lstrip("/")
+        path_parts = list(filter(lambda s: bool(s), path.split("/")))
+
+        def _matching_starts(file_path):
+            file_parts = filter(lambda s: bool(s), file_path.split("/"))
+            return all(a == b for a, b in zip(path_parts, file_parts))
+
+        self._get_dirs()
+
+        result = {}
+        # To match posix find, if an exact file name is given, we should
+        # return only that file
+        if path in self.dir_cache and self.dir_cache[path]["type"] == "file":
+            result[path] = self.dir_cache[path]
+            return result if detail else [path]
+
+        for file_path, file_info in self.dir_cache.items():
+            if not (path == "" or _matching_starts(file_path)):
+                continue
+
+            if file_info["type"] == "directory":
+                if withdirs:
+                    if file_path not in result:
+                        result[file_path.strip("/")] = file_info
+                continue
+
+            if file_path not in result:
+                result[file_path] = file_info if detail else None
+
+        if maxdepth:
+            path_depth = path.count("/")
+            result = {
+                k: v for k, v in result.items() if k.count("/") - path_depth < maxdepth
+            }
+        return result if detail else sorted(result)
