@@ -4,6 +4,7 @@ import contextlib
 import os
 import pickle
 import tempfile
+from unittest.mock import Mock
 
 import pytest
 
@@ -167,14 +168,14 @@ def test_recursive_get_put(tmpdir, m):
     fs.get("test/", d, recursive=True)
     for file in ["one", "two", "nest/other"]:
         with open(f"{d}/{file}", "rb") as f:
-            f.read() == b"data"
+            assert f.read() == b"data"
 
     # get to directory without slash
     d = tempfile.mkdtemp()
     fs.get("test", d, recursive=True)
     for file in ["test/one", "test/two", "test/nest/other"]:
         with open(f"{d}/{file}", "rb") as f:
-            f.read() == b"data"
+            assert f.read() == b"data"
 
 
 def test_pipe_cat(m):
@@ -190,7 +191,7 @@ def test_pipe_cat(m):
 def test_read_block_delimiter(m):
     fs = MemoryFileSystem()
     with fs.open("/myfile", "wb") as f:
-        f.write(b"some\n" b"lines\n" b"of\n" b"text")
+        f.write(b"some\nlines\nof\ntext")
     assert fs.read_block("/myfile", 0, 2, b"\n") == b"some\n"
     assert fs.read_block("/myfile", 2, 6, b"\n") == b"lines\n"
     assert fs.read_block("/myfile", 6, 2, b"\n") == b""
@@ -202,7 +203,7 @@ def test_read_block_delimiter(m):
 def test_open_text(m):
     fs = MemoryFileSystem()
     with fs.open("/myfile", "wb") as f:
-        f.write(b"some\n" b"lines\n" b"of\n" b"text")
+        f.write(b"some\nlines\nof\ntext")
     f = fs.open("/myfile", "r", encoding="latin1")
     assert f.encoding == "latin1"
 
@@ -308,6 +309,7 @@ def test_chained_equivalent():
     #  since the parameters don't quite match. Also, the url understood by the two
     #  of s are not the same (path gets munged a bit differently)
     assert of.fs == of2.fs
+    assert hash(of.fs) == hash(of2.fs)
     assert of.open().read() == of2.open().read()
 
 
@@ -381,50 +383,116 @@ def test_url_to_fs():
 
 
 def test_walk(m):
-    dir0 = "/dir0"
-    dir1 = dir0 + "/dir1"
-    dir2 = dir1 + "/dir2"
-    file1 = dir0 + "/file1"
-    file2 = dir1 + "/file2"
-    file3 = dir2 + "/file3"
+    # depth = 0
+    dir1 = "/dir1"
+    # depth = 1 (2 dirs, 1 file)
+    dir11 = dir1 + "/dir11"
+    dir12 = dir1 + "/dir12"
+    file11 = dir1 + "/file11"
+    # depth = 2
+    dir111 = dir11 + "/dir111"
+    file111 = dir11 + "/file111"
+    file121 = dir12 + "/file121"
+    # depth = 3
+    file1111 = dir111 + "/file1111"
 
-    m.mkdir(dir2)  # Creates parents too
-    m.touch(file1)
-    m.touch(file2)
-    m.touch(file3)
+    m.mkdir(dir111)  # Creates parents too
+    m.mkdir(dir12)  # Creates parents too
+    m.touch(file11)
+    m.touch(file111)
+    m.touch(file121)
+    m.touch(file1111)
 
     # No maxdepth
-    assert list(m.walk(dir0, topdown=True)) == [
-        (dir0, ["dir1"], ["file1"]),
-        (dir1, ["dir2"], ["file2"]),
-        (dir2, [], ["file3"]),
+    assert list(m.walk(dir1, topdown=True)) == [
+        (dir1, ["dir11", "dir12"], ["file11"]),
+        (dir11, ["dir111"], ["file111"]),
+        (dir111, [], ["file1111"]),
+        (dir12, [], ["file121"]),
     ]
-    assert list(m.walk(dir0, topdown=False)) == [
-        (dir2, [], ["file3"]),
-        (dir1, ["dir2"], ["file2"]),
-        (dir0, ["dir1"], ["file1"]),
+    assert list(m.walk(dir1, topdown=False)) == [
+        (dir111, [], ["file1111"]),
+        (dir11, ["dir111"], ["file111"]),
+        (dir12, [], ["file121"]),
+        (dir1, ["dir11", "dir12"], ["file11"]),
     ]
 
     # maxdepth=2
-    assert list(m.walk(dir0, maxdepth=2, topdown=True)) == [
-        (dir0, ["dir1"], ["file1"]),
-        (dir1, ["dir2"], ["file2"]),
+    assert list(m.walk(dir1, maxdepth=2, topdown=True)) == [
+        (dir1, ["dir11", "dir12"], ["file11"]),
+        (dir11, ["dir111"], ["file111"]),
+        (dir12, [], ["file121"]),
     ]
-    assert list(m.walk(dir0, maxdepth=2, topdown=False)) == [
-        (dir1, ["dir2"], ["file2"]),
-        (dir0, ["dir1"], ["file1"]),
+    assert list(m.walk(dir1, maxdepth=2, topdown=False)) == [
+        (dir11, ["dir111"], ["file111"]),
+        (dir12, [], ["file121"]),
+        (dir1, ["dir11", "dir12"], ["file11"]),
     ]
 
     # maxdepth=1
-    assert list(m.walk(dir0, maxdepth=1, topdown=True)) == [
-        (dir0, ["dir1"], ["file1"]),
+    assert list(m.walk(dir1, maxdepth=1, topdown=True)) == [
+        (dir1, ["dir11", "dir12"], ["file11"]),
     ]
-    assert list(m.walk(dir0, maxdepth=1, topdown=False)) == [
-        (dir0, ["dir1"], ["file1"]),
+    assert list(m.walk(dir1, maxdepth=1, topdown=False)) == [
+        (dir1, ["dir11", "dir12"], ["file11"]),
     ]
 
     # maxdepth=0
     with pytest.raises(ValueError):
-        list(m.walk(dir0, maxdepth=0, topdown=True))
+        list(m.walk(dir1, maxdepth=0, topdown=True))
     with pytest.raises(ValueError):
-        list(m.walk(dir0, maxdepth=0, topdown=False))
+        list(m.walk(dir1, maxdepth=0, topdown=False))
+
+    # prune dir111
+    def _walk(*args, **kwargs):
+        for path, dirs, files in m.walk(*args, **kwargs):
+            yield (path, dirs.copy(), files)
+            if "dir111" in dirs:
+                dirs.remove("dir111")
+
+    assert list(_walk(dir1, topdown=True)) == [
+        (dir1, ["dir11", "dir12"], ["file11"]),
+        (dir11, ["dir111"], ["file111"]),
+        (dir12, [], ["file121"]),
+    ]
+    assert list(_walk(dir1, topdown=False)) == [
+        (dir111, [], ["file1111"]),
+        (dir11, ["dir111"], ["file111"]),
+        (dir12, [], ["file121"]),
+        (dir1, ["dir11", "dir12"], ["file11"]),
+    ]
+
+    # reverse dirs order
+    def _walk(*args, **kwargs):
+        for path, dirs, files in m.walk(*args, **kwargs):
+            yield (path, dirs.copy(), files)
+            dirs.reverse()
+
+    assert list(_walk(dir1, topdown=True)) == [
+        (dir1, ["dir11", "dir12"], ["file11"]),
+        # Here dir12 comes before dir11
+        (dir12, [], ["file121"]),
+        (dir11, ["dir111"], ["file111"]),
+        (dir111, [], ["file1111"]),
+    ]
+    assert list(_walk(dir1, topdown=False)) == [
+        (dir111, [], ["file1111"]),
+        (dir11, ["dir111"], ["file111"]),
+        (dir12, [], ["file121"]),
+        (dir1, ["dir11", "dir12"], ["file11"]),
+    ]
+
+    # on_error omit by default
+    assert list(m.walk("do_not_exist")) == []
+    # on_error omit
+    assert list(m.walk("do_not_exist", on_error="omit")) == []
+    # on_error raise
+    with pytest.raises(FileNotFoundError):
+        list(m.walk("do_not_exist", on_error="raise"))
+    # on_error callable function
+    mock = Mock()
+    assert list(m.walk("do_not_exist", on_error=mock.onerror)) == []
+    mock.onerror.assert_called()
+    assert mock.onerror.call_args.kwargs == {}
+    assert len(mock.onerror.call_args.args) == 1
+    assert isinstance(mock.onerror.call_args.args[0], FileNotFoundError)
