@@ -3,6 +3,7 @@ import pickle
 import tempfile
 import zipfile
 from contextlib import contextmanager
+from pathlib import Path
 
 import pytest
 
@@ -13,13 +14,15 @@ from fsspec.core import (
     _expand_paths,
     expand_paths_if_needed,
     get_compression,
+    get_fs_token_paths,
     open_files,
     open_local,
 )
 
 
 @contextmanager
-def tempzip(data={}):
+def tempzip(data=None):
+    data = data or {}
     f = tempfile.mkstemp(suffix="zip")[1]
     with zipfile.ZipFile(f, mode="w") as z:
         for k, v in data.items():
@@ -29,7 +32,7 @@ def tempzip(data={}):
     finally:
         try:
             os.remove(f)
-        except (IOError, OSError):
+        except OSError:
             pass
 
 
@@ -57,7 +60,6 @@ def test_expand_paths(path, name_function, num, out):
     ],
 )
 def test_expand_paths_if_needed_in_read_mode(create_files, path, out):
-
     d = str(tempfile.mkdtemp())
     for f in create_files:
         f = os.path.join(d, f)
@@ -73,6 +75,11 @@ def test_expand_paths_if_needed_in_read_mode(create_files, path, out):
 def test_expand_error():
     with pytest.raises(ValueError):
         _expand_paths("*.*", None, 1)
+
+
+@pytest.mark.parametrize("mode", ["w", "w+", "x", "x+"])
+def test_expand_fs_token_paths(mode):
+    assert len(get_fs_token_paths("path", mode, num=2, expand=True)[-1]) == 2
 
 
 def test_openfile_api(m):
@@ -95,15 +102,54 @@ def test_openfile_open(m):
     assert m.size("somepath") == 5
 
 
-def test_open_local():
+def test_open_local_w_cache():
     d1 = str(tempfile.mkdtemp())
     f1 = os.path.join(d1, "f1")
     open(f1, "w").write("test1")
     d2 = str(tempfile.mkdtemp())
-    fn = open_local("simplecache://" + f1, cache_storage=d2, target_protocol="file")
+    fn = open_local(f"simplecache://{f1}", cache_storage=d2, target_protocol="file")
     assert isinstance(fn, str)
     assert open(fn).read() == "test1"
     assert d2 in fn
+
+
+def test_open_local_w_magic():
+    d1 = str(tempfile.mkdtemp())
+    f1 = os.path.join(d1, "f1")
+    open(f1, "w").write("test1")
+    fn = open_local(os.path.join(d1, "f*"))
+    assert len(fn) == 1
+    assert isinstance(fn, list)
+
+
+def test_open_local_w_list_of_str():
+    d1 = str(tempfile.mkdtemp())
+    f1 = os.path.join(d1, "f1")
+    open(f1, "w").write("test1")
+    fn = open_local([f1, f1])
+    assert len(fn) == 2
+    assert isinstance(fn, list)
+    assert all(isinstance(elem, str) for elem in fn)
+
+
+def test_open_local_w_path():
+    d1 = str(tempfile.mkdtemp())
+    f1 = os.path.join(d1, "f1")
+    open(f1, "w").write("test1")
+    p = Path(f1)
+    fn = open_local(p)
+    assert isinstance(fn, str)
+
+
+def test_open_local_w_list_of_path():
+    d1 = str(tempfile.mkdtemp())
+    f1 = os.path.join(d1, "f1")
+    open(f1, "w").write("test1")
+    p = Path(f1)
+    fn = open_local([p, p])
+    assert len(fn) == 2
+    assert isinstance(fn, list)
+    assert all(isinstance(elem, str) for elem in fn)
 
 
 def test_xz_lzma_compressions():
@@ -121,6 +167,18 @@ def test_list():
     of = open_files(plist)
     assert len(of) == len(flist)
     assert [f.path for f in of] == plist
+
+
+def test_open_expand(m, monkeypatch):
+    m.pipe("/myfile", b"hello")
+    with pytest.raises(FileNotFoundError, match="expand=True"):
+        with fsspec.open("memory://my*", expand=False):
+            pass
+    with fsspec.open("memory://my*", expand=True) as f:
+        assert f.path == "/myfile"
+    monkeypatch.setattr(fsspec.core, "DEFAULT_EXPAND", True)
+    with fsspec.open("memory://my*") as f:
+        assert f.path == "/myfile"
 
 
 def test_pathobject(tmpdir):
@@ -144,10 +202,10 @@ def test_pathobject(tmpdir):
 
 def test_automkdir(tmpdir):
     dir = os.path.join(str(tmpdir), "a")
-    of = fsspec.open(os.path.join(dir, "afile"), "w")
-    with of:
-        pass
-    assert "afile" in os.listdir(dir)
+    of = fsspec.open(os.path.join(dir, "afile"), "w", auto_mkdir=False)
+    with pytest.raises(IOError):
+        with of:
+            pass
 
     dir = os.path.join(str(tmpdir), "b")
     of = fsspec.open(os.path.join(dir, "bfile"), "w", auto_mkdir=True)
@@ -193,7 +251,110 @@ def test_pickle_after_open_open():
     of2.close()
 
 
+# Define a list of special glob characters.
+# Note that we need to escape some characters and also consider file system limitations.
+# '*' and '?' are excluded because they are not valid for many file systems.
+# Similarly, we're careful with '{', '}', and '@' as their special meaning is
+# context-specific and might not be considered special for filenames.
+# Add tests for more file systems and for more glob magic later
+glob_magic_characters = ["[", "]", "!"]
+if os.name != "nt":
+    glob_magic_characters.extend(("*", "?"))  # not valid on Windows
+
+
+@pytest.mark.parametrize("char", glob_magic_characters)
+def test_open_file_read_with_special_characters(tmp_path, char):
+    # Create a filename incorporating the special character
+    file_name = f"test{char}.txt"
+    file_path = tmp_path / file_name
+    expected_content = "Hello, world!"
+
+    with open(file_path, "w") as f:
+        f.write(expected_content)
+
+    with fsspec.open(file_path, "r") as f:
+        actual_content = f.read()
+
+    assert actual_content == expected_content
+
+
+@pytest.mark.parametrize("char", glob_magic_characters)
+def test_open_files_read_with_special_characters(tmp_path, char):
+    # Create a filename incorporating the special character
+    file_name = f"test{char}.txt"
+    file_path = tmp_path / file_name
+    expected_content = "Hello, world!"
+
+    with open(file_path, "w") as f:
+        f.write(expected_content)
+
+    with fsspec.open_files(file_path, "r")[0] as f:
+        actual_content = f.read()
+
+    assert actual_content == expected_content
+
+
+@pytest.mark.parametrize("char", glob_magic_characters)
+def test_open_file_write_with_special_characters(tmp_path, char, monkeypatch):
+    # Create a filename incorporating the special character
+    file_name = f"test{char}.txt"
+    file_path = tmp_path / file_name
+    expected_content = "Hello, world!"
+
+    with fsspec.open(file_path, "w", expand=False) as f:
+        f.write(expected_content)
+
+    with open(file_path, "r") as f:
+        actual_content = f.read()
+
+    monkeypatch.setattr(fsspec.core, "DEFAULT_EXPAND", False)
+    with fsspec.open(file_path, "w") as f:
+        f.write(expected_content * 2)
+
+    with open(file_path, "r") as f:
+        assert f.read() == actual_content * 2
+
+    assert actual_content == expected_content
+
+
+@pytest.mark.parametrize("char", glob_magic_characters)
+def test_open_files_read_with_special_characters(tmp_path, char):
+    # Create a filename incorporating the special character
+    file_name = f"test{char}.txt"
+    file_path = tmp_path / file_name
+    expected_content = "Hello, world!"
+
+    with open(file_path, "w") as f:
+        f.write(expected_content)
+
+    with fsspec.open_files(
+        urlpath=[os.fspath(file_path)], mode="r", auto_mkdir=False, expand=False
+    )[0] as f:
+        actual_content = f.read()
+
+    assert actual_content == expected_content
+
+
+@pytest.mark.parametrize("char", glob_magic_characters)
+def test_open_files_write_with_special_characters(tmp_path, char):
+    # Create a filename incorporating the special character
+    file_name = f"test{char}.txt"
+    file_path = tmp_path / file_name
+    expected_content = "Hello, world!"
+
+    with fsspec.open_files(
+        urlpath=[os.fspath(file_path)], mode="w", auto_mkdir=False, expand=False
+    )[0] as f:
+        f.write(expected_content)
+
+    with open(file_path, "r") as f:
+        actual_content = f.read()
+
+    assert actual_content == expected_content
+
+
 def test_mismatch():
+    pytest.importorskip("s3fs")
     with pytest.raises(ValueError):
         open_files(["s3://test/path.csv", "/other/path.csv"])
 
@@ -299,3 +460,26 @@ def test_chained_url(ftp_writable):
         url += f"::ftp://{username}:{password}@{host}:{port}/archive.zip"
         with fsspec.open(url, "rb") as f:
             assert f.read() == data["afile"]
+
+
+def test_automkdir_local():
+    fs, _ = fsspec.core.url_to_fs("file://", auto_mkdir=True)
+    assert fs.auto_mkdir is True
+
+
+def test_repeated_argument():
+    pytest.importorskip("adlfs")
+    from fsspec.core import url_to_fs
+
+    fs, url = url_to_fs(
+        "az://DIR@ACCOUNT.blob.core.windows.net/DATA",
+        anon=False,
+        account_name="ACCOUNT",
+    )
+    assert fs.storage_options == {"account_name": "ACCOUNT", "anon": False}
+    with pytest.raises(TypeError):
+        url_to_fs(
+            "az://DIR@ACCOUNT.blob.core.windows.net/DATA",
+            anon=False,
+            account_name="OTHER",
+        )
