@@ -42,7 +42,6 @@ def engine(request):
 def test_open_parquet_file(
     tmpdir, engine, columns, max_gap, max_block, footer_sample_size, range_index
 ):
-
     # Pandas required for this test
     pd = pytest.importorskip("pandas")
 
@@ -92,7 +91,7 @@ def test_open_parquet_file(
             # the footer metadata, so there should NOT
             # be a key for the last 8 bytes of the file
             bad_key = (file_size - 8, file_size)
-            assert bad_key not in data.keys()
+            assert bad_key not in data
 
         for (start, stop), byte_data in data.items():
             f.seek(start)
