@@ -1,8 +1,11 @@
 import sys
+from importlib.metadata import EntryPoint
 from unittest.mock import create_autospec, patch
 
 import pytest
 
+import fsspec
+from fsspec.implementations.zip import ZipFileSystem
 from fsspec.registry import (
     _registry,
     filesystem,
@@ -12,11 +15,6 @@ from fsspec.registry import (
     registry,
 )
 from fsspec.spec import AbstractFileSystem
-
-try:
-    from importlib.metadata import EntryPoint
-except ImportError:  # python < 3.8
-    from importlib_metadata import EntryPoint
 
 
 @pytest.fixture()
@@ -73,9 +71,16 @@ def test_register_fail(clear_registry):
     with pytest.raises(ImportError):
         get_filesystem_class("test")
 
-    register_implementation("test", "doesntexist.AbstractFileSystem")
+    # NOOP
+    register_implementation("test", "doesntexist.AbstractFileSystem", clobber=False)
     with pytest.raises(ValueError):
-        register_implementation("test", "doesntexist.AbstractFileSystem", clobber=False)
+        register_implementation(
+            "test", "doesntexist.AbstractFileSystemm", clobber=False
+        )
+
+    # by default we do not allow clobbering
+    with pytest.raises(ValueError):
+        register_implementation("test", "doesntexist.AbstractFileSystemm")
 
     register_implementation(
         "test", "doesntexist.AbstractFileSystem", errtxt="hiho", clobber=True
@@ -85,22 +90,22 @@ def test_register_fail(clear_registry):
     assert "hiho" in str(e.value)
     register_implementation("test", AbstractFileSystem)
 
+    # NOOP
+    register_implementation("test", AbstractFileSystem)
     with pytest.raises(ValueError):
-        register_implementation("test", AbstractFileSystem, clobber=False)
+        register_implementation("test", ZipFileSystem)
     register_implementation("test", AbstractFileSystem, clobber=True)
+    assert isinstance(fsspec.filesystem("test"), AbstractFileSystem)
 
 
 def test_entry_points_registered_on_import(clear_registry, clean_imports):
     mock_ep = create_autospec(EntryPoint, module="fsspec.spec.AbstractFileSystem")
     mock_ep.name = "test"  # this can't be set in the constructor...
     mock_ep.value = "fsspec.spec.AbstractFileSystem"
-    if sys.version_info < (3, 8):
-        import_location = "importlib_metadata.entry_points"
-    else:
-        import_location = "importlib.metadata.entry_points"
+    import_location = "importlib.metadata.entry_points"
     with patch(import_location, return_value={"fsspec.specs": [mock_ep]}):
         assert "test" not in registry
-        import fsspec  # noqa
+        import fsspec  # noqa: F401
 
         get_filesystem_class("test")
         assert "test" in registry
@@ -110,12 +115,20 @@ def test_filesystem_warning_arrow_hdfs_deprecated(clear_registry, clean_imports)
     mock_ep = create_autospec(EntryPoint, module="fsspec.spec.AbstractFileSystem")
     mock_ep.name = "arrow_hdfs"  # this can't be set in the constructor...
     mock_ep.value = "fsspec.spec.AbstractFileSystem"
-    if sys.version_info < (3, 8):
-        import_location = "importlib_metadata.entry_points"
-    else:
-        import_location = "importlib.metadata.entry_points"
+    import_location = "importlib.metadata.entry_points"
     with patch(import_location, return_value={"fsspec.specs": [mock_ep]}):
-        import fsspec  # noqa
+        import fsspec  # noqa: F401
 
         with pytest.warns(DeprecationWarning):
             filesystem("arrow_hdfs")
+
+
+def test_old_s3(monkeypatch):
+    from fsspec.registry import _import_class
+
+    s3fs = pytest.importorskip("s3fs")
+    monkeypatch.setattr(s3fs, "__version__", "0.4.2")
+    with pytest.warns():
+        _import_class("s3fs:S3FileSystem")
+    with pytest.warns():
+        _import_class("s3fs.S3FileSystem")
